+++ material/src/material.ts
@@ -1,3 +1,8 @@
 export {TitledDialogModule} from './modules/titledDialog.module';
 export {TitledDialogService} from './services/titledDialog/titledDialog.service';
-export {TitledDialogConfig, TITLED_DIALOG_DATA} from './misc/titledDialog.interface';
+export {TitledDialogConfig, TITLED_DIALOG_DATA} from './misc/interfaces/titledDialog.interface';
+export {ConfirmationDialogModule} from './modules/confirmationDialog.module';
+export {ConfirmationDialogComponent} from './components/confirmationDialog/confirmationDialog.component';
+export {ConfirmationDialogDirective} from './directives/confirmationDialog/confirmationDialog.directive';
+export {ConfirmationDialogData} from './misc/interfaces/confirmationDialog.interface';
+export {CONFIRMATION_DIALOG_DATA} from './misc/tokens';