+++ material/src/services/titledDialog/titledDialog.service.ts
@@ -2,7 +2,7 @@ import {Injectable, TemplateRef} from "@angular/core";
 import {MatDialog, MatDialogRef} from "@angular/material/dialog";
 import {ComponentType} from "@angular/cdk/portal";
 
-import {TitledDialogConfig, ɵTitledDialogOptions} from "../../misc/titledDialog.interface";
+import {TitledDialogConfig, ɵTitledDialogOptions} from "../../misc/interfaces/titledDialog.interface";
 import {TitledDialogComponent} from "../../components/titledDialog/titledDialog.component";
 
 /**